add wrapper sub around inner calls to after_import
The core extensions were using the caller level in after_import to find
the importing package. They have been adjusted to instead be passed in
that value and be able to use it directly. But some other subclasses may
still try to use the caller. Add a wrapping sub so that it matches the
old number of stack frames to find the importing package.