fix log4perl test when run with unexpected paths
Fix the log4perl test to expect the file path that the test was actually
loaded with. This allows it to work with whatever was used to load the
file, rather than just the standard path that would be used on unix
systems. This allows it to work with alternate directory separators.