remove version number from pod in source, leaving placeholder
The $VERSION in the source code gets updated during the release process.
However, the current process is not able to update the version in the
Pod in the source. It does get correctly replaced in the generated
tarball.

Since it doesn't get updated, it will fall out of date unless manually
updated. Rather than deal with that, just remove the version. Leave a
placeholder, so that the line numbers of the remaining Pod doesn't
change between source and release.